Setup basic animations
Configured animations for jogging in different directions including strafing.
Added the ability to shoot the weapon with recoil and particle effects for muzzle burst, smoke trail, and hit impact. Insured with a trace that the impact point is correct if a blocking object is in front of the target destination.

Leveraging Balica from Paragon only animations that were edited are included in the commit, the others need to add from the marketplace seperately.