Setup Enemy Class
Created a Enemy C++ class to handle health and hits. Added to support for headshots and body shots. Created a widget to display the hit amounts and set a different color for when it was a successful head shot. Updated the weapons in the DT to have there own body/head shot values.